End project

+++ routers/client.js
@@ -1,5 +1,6 @@
 const express = require("express");
 const { db } = require("../utils/db");
+const { NotFoundError } = require("../utils/errors");
 
 const clientRouter = express.Router();
 
@@ -11,6 +12,12 @@ clientRouter
   })
 
   .get("/:id", (req, res) => {
+    const client = db.getOne(req.params.id);
+
+    if (!client) {
+      throw new NotFoundError();
+    }
+
     res.render("client/one", {
       clients: db.getOne(req.params.id),
     });
@@ -19,13 +26,19 @@ clientRouter
   .post("/", (req, res) => {
     const id = db.create(req.body);
 
-    res.render("client/added", {
+    res.status(201).render("client/added", {
       name: req.body.name,
       id,
     });
   })
 
   .put("/:id", (req, res) => {
+    const client = db.getOne(req.params.id);
+
+    if (!client) {
+      throw new NotFoundError();
+    }
+
     db.update(req.params.id, req.body);
     res.render("client/modified", {
       name: req.body.name,
@@ -34,6 +47,12 @@ clientRouter
   })
 
   .delete("/:id", (req, res) => {
+    const client = db.getOne(req.params.id);
+
+    if (!client) {
+      throw new NotFoundError();
+    }
+
     db.delete(req.params.id);
     res.render("client/deleted");
   })
@@ -43,6 +62,12 @@ clientRouter
   })
 
   .get("/form/edit/:id", (req, res) => {
+    const client = db.getOne(req.params.id);
+
+    if (!client) {
+      throw new NotFoundError();
+    }
+
     res.render("client/forms/add", {
       client: db.getOne(req.params.id),
     });
